create mixin to the get_price method, add display offers and dlcs for GameDetail
