+++ backend/games/finance/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
-from finance.views import OfferAPIView, ShowLibraryView, ShowProductContentView
+from finance.views import OfferAPIView, ShowLibraryView
 
 
 urlpatterns = [
     path('offer/', OfferAPIView.as_view(), name="offer"),
-    path("library/<user>", ShowLibraryView.as_view(), name="all_products_in_library"),
-    path('pack/<uuid:pk>/', ShowProductContentView.as_view(), name='all_content_for_product')
+    path("library/<user>", ShowLibraryView.as_view(), name="all_products_in_library")
 ]
