+++ backend/games/finance/views.py
@@ -1,9 +1,8 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework import generics
-from core.models import Product
 from finance.models import Library
-from finance.serializers import OfferSerializer, LibrarySerializer, ProductContentShowSerializer
+from finance.serializers import OfferSerializer, LibrarySerializer
 
 
 class OfferAPIView(APIView):
@@ -18,10 +17,3 @@ class ShowLibraryView(generics.RetrieveAPIView):
     serializer_class = LibrarySerializer
     queryset = Library.objects.all()
     lookup_field = 'user'
-
-
-class ShowProductContentView(generics.RetrieveAPIView):
-    serializer_class = ProductContentShowSerializer
-
-    def get_queryset(self):
-        return Product.objects.all()
